backend/pyproject.toml
[tool.poetry]
name = "tome"
version = "0.0.1"
description = "A notebook and wiki app built for how you think"
authors = ["pxeger <_@pxeger.com>"]

[tool.poetry.dependencies]
python = "^3.8"
starlette = "^0.13.4"
asyncpg = "^0.20.1"
aiosmtplib = "^1.1.3"
orjson = "^3.2.0"
pyjwt = "^1.7.1"
argon2-cffi = "^20.1.0"
jinja2 = "^2.11.2"

[tool.poetry.dev-dependencies]
flakehell = "^0.5.0"
mypy = "^0.782"
black = "^19.10b0"
hacking = "^3.1.0"
flake8-bugbear = "^20.1.4"
isort = "^4.3.21"
pytest = "^5.4.3"
uvicorn = "^0.11.5"
httpx = "^0.13.3"
asgi-lifespan = "^1.0.1"
pytest-asyncio = "^0.14.0"

[tool.flakehell]
max_line_length = 88

[tool.flakehell.plugins]
"*" = ["+*", "-H203", "-H204", "-H205", "-H210", "-H904", "-H301", "-H306", "-E501", "-W503"]
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

